in_tail: use inode for key of TailWatchers
This allows unwatching to be completely handled in `refresh_wathcer`.

`update_watcher` needed to unwatch because the key was path, but it
caused log duplication in some cases.

This key change fixes that problem.

Note about the key of TailWatchers:
Regardless of `follow_inodes`, both path or inode is acceptable
for the key of `tails`.
It is not wrong that the current logic uses `path` for the key.
It just means that using inode as the key can solve the unwatching
problem easily.

perhaps we don't need to use different keys depending on
`follow_inodes`. We can always use inode for the key. (Future work)

Signed-off-by: Katuya Kawakami <kkawakami-je@nec.com>
Signed-off-by: Masaki Hatada <mas-hatada@nec.com>
Co-authored-by: Daijiro Fukuda <fukuda@clear-code.com>