in_tail: prevent wrongly unwatching with follow_inodes
We must not unwatch targets that still exist.
If unwatching an existing target, it causes log duplication.

In the existing implementation, `update_watcher` needs to unwatch
the old TailWatcher since `refresh_watcher` can't unwatch it when
`update_watcher` is called first.
It is because `update_watcher` discards the old TailWatcher and
`refresh_watcher` can't recognize the old inode is disappeared.

However, it can wrongly unwatch an existing inode because the
old inode may still exist.
(See the diff of test cases.)

Thus, we need a new mechanism to correctly unwatch targets.

This fix is based on the idea that we should unwatch based on
directly on PositionFile's data.

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>
Co-authored-by: Takuro Ashie <ashie@clear-code.com>